scripts/fmi/fmu/DTMockup/resources/requirements.txt
arrow==1.3.0
attrs==23.1.0
click==7.1.2
contourpy==1.2.0
cycler==0.12.1
docopt==0.6.2
FMPy==0.3.19
fonttools==4.46.0
importlib-metadata==4.13.0
importlib-resources==6.1.1
Jinja2==3.1.2
kiwisolver==1.4.5
lark==1.1.8
loguru==0.7.2
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.4
mosaik==3.2.0
mosaik-api-v3==3.0.5
mosaik-csv==2.0.0
msgpack==1.0.7
networkx==3.2.1
numpy
packaging==23.2
pandas==1.3.5
pathlib==1.0.1
Pillow==10.1.0
Pint==0.20.1
plotly==5.18.0
# pylzma==0.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pythonfmu==0.6.3
pytz==2023.3.post1
PyYAML==5.3.1
rainflow==3.2.0
scipy==1.9.3
six==1.16.0
tenacity==8.2.3
tqdm==4.65.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
zipp==3.17.0
